+++ src/REvoDesign/common/file_extensions.py
@@ -29,9 +29,20 @@ PDB = ExtColl(
         Ext("mmcif", "Macromolecular Crystallographic Information File"),
     )
 )
+PDB_STRICT = ExtColl(
+    (
+        Ext("pdb", "Protein Data Bank format file"),
+    )
+)
 MOL = ExtColl(
     (
         Ext("mol", "Mol2 file"),
+        Ext("sdf", "SDF file"),
+    )
+)
+
+SDF = ExtColl(
+    (
         Ext("sdf", "SDF file"),
     )
 )
@@ -85,5 +96,11 @@ YAML = ExtColl(
 JSON = ExtColl(
     (
         Ext("json", "JSON file"),
+    )
+)
+
+RosettaParams = ExtColl(
+    (
+        Ext("params", "Rosetta Parameter file"),
     )
 )