+++ src/REvoDesign/driver/file_dialog.py
@@ -3,14 +3,15 @@ File Dialog
 '''
 import os
 from functools import partial
-from typing import Any, Literal, Optional
+from typing import Any, List, Literal, Optional
 
 from pymol.Qt.utils import getSaveFileNameWithExt
 
 from ..basic import FileExtensionCollection, SingletonAbstract
 from ..common import file_extensions
 from ..logger import ROOT_LOGGER
-from ..tools.customized_widgets import decide, getOpenFileNameWithExt
+from ..tools.customized_widgets import (decide, getMultipleFiles,
+                                        getOpenFileNameWithExt)
 from ..tools.utils import extract_archive
 from .ui_driver import ConfigBus
 
@@ -45,6 +46,10 @@ class FileDialog(SingletonAbstract):
 
     # class public function that can be shared with each tab
     # callback for the "Browse" button
+
+    def browse_multiple_files(self, exts: Optional[tuple[FileExtensionCollection, ...]] = (file_extensions.Any,)
+                              ) -> List[str]:
+        return getMultipleFiles(self.window, exts)
 
     def browse_filename(
         self, mode: IO_MODE = "r", exts: tuple[FileExtensionCollection, ...] = (file_extensions.Any,)